working commit

+++ defaults.go
@@ -6,7 +6,7 @@ var (
 	pongWait       = 60 * time.Second
 	pingPeriod     = (pongWait * 9) / 10
 	writeWait      = 10 * time.Second
-	maxMessageSize = uint32(512)
+	maxMessageSize = int64(512)
 	maxReadBufferSize = 1024
 	maxWriteBufferSize = 1024
 )
